BaytesOptim

Documentation for BaytesOptim.

BaytesOptim.ConfigLBFG

BaytesOptim.ConfigSGD

BaytesOptim.DiagnosticsLBFG

BaytesOptim.DiagnosticsSGD

BaytesOptim.LBFGTune

BaytesOptim.OptimConstructor

BaytesOptim.OptimDefault

BaytesOptim.OptimTune

BaytesOptim.Optimizer

BaytesOptim.SGDTune

BaytesCore.generate

BaytesCore.generate_showvalues

BaytesCore.generate_showvalues

BaytesCore.infer

BaytesCore.infer

BaytesCore.init

BaytesCore.init

BaytesCore.init

BaytesCore.init

BaytesCore.propagate

BaytesCore.propagate

BaytesCore.propose

BaytesCore.propose!

BaytesCore.results

BaytesCore.update!

BaytesOptim._optimize

BaytesOptim.infer_generated

BaytesOptim.ConfigLBFG — Type

struct ConfigLBFG{R<:Real} <: BaytesCore.AbstractConfiguration

Default Configuration for LBFG optimizer.

Fields

magnitude_penalty::Real: Add -0.5 * magnitude_penalty * sum(abs2, q) to the log posterior when finding the local optimum. This can help avoid getting into high-density edge areas of the posterior which are otherwise not typical (eg multilevel models).

iterations::Int64: Maximum number of iterations in the optimization algorithm. Recall that we don't need to find the mode, or even a local mode, just be in a reasonable region.

difforder::BaytesDiff.DiffOrderOne: Differentiable order for objective function needed to run propagate step

BaytesOptim.ConfigSGD — Type

struct ConfigSGD{R<:Real, A<:UpdateBool} <: BaytesCore.AbstractConfiguration

Default Configuration for SGD optimizer.

Fields

magnitude_penalty::Real: Add -0.5 * magnitude_penalty * sum(abs2, q) to the log posterior when finding the local optimum. This can help avoid getting into high-density edge areas of the posterior which are otherwise not typical (eg multilevel models).

magnitude_adaption::UpdateBool: Adapt magnitude iteratively for each step ~ currently not implemented

iterations::Int64: Maximum number of iterations in the optimization algorithm. Recall that we don't need to find the mode, or even a local mode, just be in a reasonable region.

difforder::BaytesDiff.DiffOrderOne: Differentiable order for objective function needed to run propagate step

BaytesOptim.DiagnosticsLBFG — Type

struct DiagnosticsLBFG <: OptimKernelDiagnostics

Diagnostics for LBFG.

Fields

BaytesOptim.DiagnosticsSGD — Type

struct DiagnosticsSGD <: OptimKernelDiagnostics

Diagnostics for SGD.

Fields

BaytesOptim.LBFGTune — Type

struct LBFGTune{R<:Real} <: BaytesCore.AbstractTune

Stores information used throughout optimization algorithms.

Fields

magnitude_penalty::Real: Add -0.5 * magnitude_penalty * sum(abs2, q) to the log posterior when finding the local optimum. This can help avoid getting into high-density edge areas of the posterior which are otherwise not typical (eg multilevel models).

iterations::Int64: Maximum number of iterations in the optimization algorithm. Recall that we don't need to find the mode, or even a local mode, just be in a reasonable region.

BaytesOptim.OptimConstructor — Type

Callable struct to make initializing Optimizer easier in sampling library.

Examples

BaytesOptim.OptimDefault — Type

struct OptimDefault{T<:NamedTuple, G, I<:AbstractInitialization, U<:UpdateBool}

Default arguments for Optim constructor.

Fields

kernel::NamedTuple: Tuning Arguments for individual Optimizer

GradientBackend::Any: Gradient backend used in Optimization step. Not used if Metropolis sampler is chosen.

init::AbstractInitialization: Method to obtain initial Modelparameter, see 'AbstractInitialization'.

generated::UpdateBool: Boolean if generate(_rng, objective) for corresponding model is stored in Optimization Diagnostics.

BaytesOptim.OptimTune — Type

struct OptimTune{T<:ModelWrappers.Tagged, K, B<:UpdateBool} <: BaytesCore.AbstractTune

Stores information used throughout optimization algorithms.

Fields

tagged::ModelWrappers.Tagged: Tagged Parameter for Optimization routine

kernel::Any: Tuning arguments for individual Optimizer

generated::UpdateBool: Boolean if generated quantities should be generated while sampling

iter::BaytesCore.Iterator: Current iteration number

BaytesOptim.Optimizer — Type

struct Optimizer{M<:OptimKernel, N<:OptimTune} <: BaytesCore.AbstractAlgorithm

Stores information for proposal step.

Fields

kernel::OptimKernel: Optimizer

tune::OptimTune: Tuning configuration for kernel.

BaytesOptim.SGDTune — Type

struct SGDTune{R<:Real, A<:UpdateBool} <: BaytesCore.AbstractTune

Stores information used throughout optimization algorithms.

Fields

magnitude_penalty::Real: Add -0.5 * magnitude_penalty * sum(abs2, q) to the log posterior when finding the local optimum. This can help avoid getting into high-density edge areas of the posterior which are otherwise not typical (eg multilevel models).

magnitude_adaption::UpdateBool: Adapt magnitude iteratively for each step ~ currently not implemented

iterations::Int64: Maximum number of iterations in the optimization algorithm. Recall that we don't need to find the mode, or even a local mode, just be in a reasonable region.

BaytesCore.generate — Method

generate(_rng, algorithm, objective)

Generate statistics for algorithm given model parameter and data.

Examples

BaytesCore.generate_showvalues — Method

generate_showvalues(diagnostics)

Show relevant diagnostic results.

Examples

BaytesCore.generate_showvalues — Method

generate_showvalues(diagnostics)

Show relevant diagnostic results.

Examples

BaytesCore.infer — Method

infer(_rng, optimizer, model, data)

Infer type of predictions of Optim sampler.

Examples

BaytesCore.infer — Method

infer(_rng, diagnostics, optimizer, model, data)

Infer Optim diagnostics type.

Examples

BaytesCore.init — Method

init(, optim, objective; magnitude_penalty, iterations)

Initialize Mala custom configurations.

Examples

BaytesCore.init — Method

init(
    ,
    optim,
    objective;
    magnitude_penalty,
    magnitude_adaption,
    iterations
)

Initialize Mala custom configurations.

Examples

BaytesCore.init — Method

init(, config, objective)

Initialize LBFG custom configurations.

Examples

BaytesCore.init — Method

init(, config, objective)

Initialize SGD custom configurations.

Examples

BaytesCore.propagate — Method

propagate(_rng, kernel, tune, objective)

Propagate forward one MALA step.

Examples

BaytesCore.propagate — Method

propagate(_rng, kernel, tune, objective)

Propagate forward one MALA step.

Examples

BaytesCore.propose! — Method

propose!(_rng, optim, model, data)
propose!(_rng, optim, model, data, proposaltune)

Inplace version of propose.

Examples

BaytesCore.propose — Method

propose(_rng, optim, objective)

Propose new parameter with optimizer. If update=true, objective function will be updated with input model and data.

Examples

BaytesCore.results — Method

results(diagnosticsᵛ, algorithm, Ndigits, quantiles)

Print result for a single trace.

Examples

BaytesCore.update! — Method

update!(tune, result)

Update Optim tuning fields at current iteration.

Examples

BaytesOptim._optimize — Function

Internal optimization function via 'Optim' and 'NLSolversBase' - both have to be loaded and are used as an Extension if used.

Examples

BaytesOptim.infer_generated — Method

infer_generated(_rng, optimizer, model, data)

Infer type of generated quantities of Optimizer.

Examples