BaytesOptim

Documentation for BaytesOptim.

BaytesOptim.LBFGTune

BaytesOptim.OptimDefault

BaytesOptim.OptimTune

BaytesOptim.Optimizer

BaytesCore.init

BaytesCore.propagate

BaytesCore.propose

BaytesCore.propose!

BaytesOptim.LBFGTune — Type

struct LBFGTune{R<:Real} <: BaytesCore.AbstractTune

Stores information used throughout optimization algorithms.

Fields

magnitude_penalty::Real: Add -0.5 * magnitude_penalty * sum(abs2, q) to the log posterior when finding the local optimum. This can help avoid getting into high-density edge areas of the posterior which are otherwise not typical (eg multilevel models).

iterations::Int64: Maximum number of iterations in the optimization algorithm. Recall that we don't need to find the mode, or even a local mode, just be in a reasonable region.

BaytesOptim.OptimDefault — Type

struct OptimDefault{T<:NamedTuple, G, I<:AbstractInitialization, U<:UpdateBool}

Default arguments for Optim constructor.

Fields

kernel::NamedTuple: Tuning Arguments for individual Optimizer

GradientBackend::Any: Gradient backend used in MCMC step. Not used if Metropolis sampler is chosen.

init::AbstractInitialization: Method to obtain initial Modelparameter, see 'AbstractInitialization'.

generated::UpdateBool: Boolean if generate(_rng, objective) for corresponding model is stored in MCMC Diagnostics.

BaytesOptim.OptimTune — Type

struct OptimTune{T<:ModelWrappers.Tagged, K} <: BaytesCore.AbstractTune

Stores information used throughout optimization algorithms.

Fields

tagged::ModelWrappers.Tagged: Tagged Parameter for Optimization routine

kernel::Any: Tuning arguments for individual Optimizer

BaytesOptim.Optimizer — Type

struct Optimizer{M<:OptimKernel, N<:OptimTune} <: BaytesCore.AbstractAlgorithm

Stores information for proposal step.

Fields

kernel::OptimKernel: MCMC sampler

tune::OptimTune: Tuning configuration for kernel.

BaytesCore.init — Method

init(, optim, objective; magnitude_penalty, iterations)

Initialize HMC custom configurations.

Examples

BaytesCore.propagate — Method

propagate(_rng, kernel, tune, objective)

Propagate forward one MALA step.

Examples

BaytesCore.propose! — Method

propose!(_rng, optim, model, data)
propose!(_rng, optim, model, data, proposaltune)

Inplace version of propose.

Examples

BaytesCore.propose — Method

propose(_rng, optim, objective)

Propose new parameter with optimizer. If update=true, objective function will be updated with input model and data.

Examples